move loading/writing of beats
