+++ bin/GMMPatternTracker
@@ -19,6 +19,7 @@ from madmom.audio.spectrogram import (FilteredSpectrogramProcessor,
 from madmom.features import ActivationsProcessor
 from madmom.features.downbeats import PatternTrackingProcessor
 from madmom.models import PATTERNS_BALLROOM
+from madmom.io import write_beats
 
 
 def main():
@@ -112,19 +113,13 @@ def main():
         # save the multiband features to file
         out_processor = ActivationsProcessor(mode='w', **vars(args))
     else:
-        # track the patterns
+        # track the (down-)beats (i.e. patterns) and output them
         out_processor = [PatternTrackingProcessor(**vars(args))]
         if args.downbeats:
             # filter the beats to contain only downbeats
             from madmom.features.downbeats import filter_downbeats
             out_processor.append(filter_downbeats)
-            # simply write the timestamps
-            from madmom.io import write_events as writer
-        else:
-            # borrow the note writer for outputting timestamps + beat numbers
-            from madmom.features.notes import write_notes as writer
-        # sequentially process them
-        out_processor.append(writer)
+        out_processor.append(write_beats)
 
     # create an IOProcessor
     processor = IOProcessor(in_processor, out_processor)
