+++ madmom/evaluation/beats.py
@@ -30,12 +30,13 @@ from __future__ import absolute_import, division, print_function
 
 from functools import wraps
 import warnings
+
 import numpy as np
 
-from . import (find_closest_matches, calc_errors, calc_absolute_errors,
-               evaluation_io, MeanEvaluation)
+from . import (MeanEvaluation, calc_absolute_errors, calc_errors,
+               evaluation_io, find_closest_matches)
 from .onsets import OnsetEvaluation
-from ..utils import suppress_warnings
+from ..io import load_beats
 
 
 # exceptions
@@ -116,53 +117,6 @@ def _score_decorator(perfect_score, zero_score):
 
 score_10 = _score_decorator(1., 0.)
 score_1100 = _score_decorator((1., 1.), (0., 0.))
-
-
-@suppress_warnings
-def load_beats(values, downbeats=False):
-    """
-    Load the beats from the given values or file.
-
-    To make this function more universal, it also accepts lists or arrays.
-
-    Parameters
-    ----------
-    values : str, file handle, list or numpy array
-        Name / values to be loaded.
-    downbeats : bool, optional
-        Load downbeats instead of beats.
-
-    Returns
-    -------
-    numpy array
-        Beats.
-
-    Notes
-    -----
-    Expected format:
-
-    'beat_time' [additional information will be ignored]
-
-    """
-    # load the beats from the given representation
-    if values is None:
-        # return an empty array
-        values = np.zeros(0)
-    elif isinstance(values, (list, np.ndarray)):
-        # convert to numpy array if possible
-        # Note: use array instead of asarray because of ndmin
-        values = np.array(values, dtype=np.float, ndmin=1, copy=False)
-    else:
-        # try to load the data from file
-        values = np.loadtxt(values, ndmin=1)
-    if values.ndim > 1:
-        if downbeats:
-            # rows with a "1" in the 2nd column are the downbeats.
-            return values[values[:, 1] == 1][:, 0]
-        else:
-            # 1st column is the beat time, the rest is ignored
-            return values[:, 0]
-    return values
 
 
 # function for sequence variations generation
