+++ madmom/io/__init__.py
@@ -115,3 +115,66 @@ def load_onsets(values):
 
 write_onsets = write_events
 
+
+@suppress_warnings
+def load_beats(values, downbeats=False):
+    """
+    Load the beats from the given values or file.
+
+    Parameters
+    ----------
+    values : str, file handle, list or numpy array
+        Name / values to be loaded.
+    downbeats : bool, optional
+        Load downbeats instead of beats.
+
+    Returns
+    -------
+    numpy array
+        Beats.
+
+    Notes
+    -----
+    Expected format:
+
+    'beat_time' ['beat_number']
+
+    """
+    # load the beats from the given representation
+    if values is None:
+        # return an empty array
+        values = np.zeros(0)
+    elif isinstance(values, (list, np.ndarray)):
+        # convert to numpy array if possible
+        # Note: use array instead of asarray because of ndmin
+        values = np.array(values, dtype=np.float, ndmin=1, copy=False)
+    else:
+        # try to load the data from file
+        values = np.loadtxt(values, ndmin=1)
+    if values.ndim > 1:
+        if downbeats:
+            # rows with a "1" in the 2nd column are the downbeats.
+            return values[values[:, 1] == 1][:, 0]
+        else:
+            # 1st column is the beat time, the rest is ignored
+            return values[:, 0]
+    return values
+
+
+def write_beats(beats, filename, **kwargs):
+    """
+    Write the beats to a file.
+
+    Parameters
+    ----------
+    beats : numpy array
+        Beats to be written to file.
+    filename : str or file handle
+        File to write the events to.
+
+    """
+    if beats.ndim == 2:
+        fmt = list(('%.3f', '%d'))
+    else:
+        fmt = '%.3f'
+    write_events(beats, filename, fmt=fmt, **kwargs)
